AE-2006: Replace toimenpide-osapuolet-table.svelte with the new implementation
- Delete notice-bailiff-osapuolet-table.svelte
