+++ src/pages/valvonta-kaytto/toimenpide-osapuolet-table.svelte
@@ -10,6 +10,12 @@
   import { _, locale } from '@Language/i18n';
   import H2 from '@Component/H/H2.svelte';
   import Spinner from '@Component/Spinner/Spinner.svelte';
+  import * as Selects from '@Component/Select/select-util';
+  import { isValid } from '@Utility/classification';
+  import Select from '@Component/Select/Select2';
+  import * as Toimenpiteet from '@Pages/valvonta-kaytto/toimenpiteet';
+  import Checkbox from '@Component/Checkbox/Checkbox';
+  import Input from '@Component/Input/Input.svelte';
 
   export let id;
   export let toimenpide;
@@ -21,8 +27,15 @@
   export let template;
   export let previewPending;
   export let disabled;
-
+  export let schema;
+  export let hallintoOikeudet = [];
+  export let karajaoikeudet = [];
   export let manuallyDeliverableToimenpide = false;
+  export let showDeliveryMethod = false;
+  export let showHallintoOikeudetSelection = false;
+  export let showKarajaOikeudetSelection = false;
+  export let showCreateDocument = false;
+  export let allowPreviewAlways = false;
 
   const types = {
     yritys: {
@@ -51,13 +64,37 @@
     Maybe.fold('', Locales.labelForId($locale, roolit)),
     R.prop('rooli-id')
   );
+
   $: toimitustapaLabel = R.compose(
     Maybe.fold('', Locales.labelForId($locale, toimitustavat)),
     R.prop('toimitustapa-id')
   );
+
+  const osapuoliSpecificDataIndexForOsapuoli = osapuoli =>
+    Toimenpiteet.osapuoliSpecificDataIndexForOsapuoli(
+      toimenpide,
+      osapuoli.id,
+      Osapuolet.getOsapuoliType(osapuoli)
+    );
+
+  const isPreviewHidden = (toimenpide, osapuoli) => {
+    if (allowPreviewAlways) {
+      return false;
+    } else {
+      return !Toimenpiteet.documentExistsForOsapuoli(
+        toimenpide,
+        osapuoli.id,
+        Osapuolet.getOsapuoliType(osapuoli)
+      );
+    }
+  };
 </script>
 
-<style>
+<style type="text/postcss">
+  .court-container {
+    height: 5rem;
+  }
+
   .etp-table--td.text-error {
     @apply font-bold;
   }
@@ -65,6 +102,7 @@
 
 <div class="w-full">
   <H2 text={i18n(i18nRoot + '.vastaanottajat')} />
+
   <div class="flex flex-row my-4">
     <table class="etp-table">
       <thead class="etp-table--thead">
@@ -73,13 +111,33 @@
 
           <th class="etp-table--th"> {i18n(i18nRoot + '.rooli')} </th>
 
-          <th class="etp-table--th">
-            {i18n(i18nRoot + '.toimitustapa')}
-          </th>
+          {#if showDeliveryMethod}
+            <th class="etp-table--th">
+              {i18n(i18nRoot + '.toimitustapa')}
+            </th>
+          {/if}
+
+          {#if showHallintoOikeudetSelection}
+            <th class="etp-table--th"> {i18n(i18nRoot + '.court')} </th>
+          {/if}
+
+          {#if showKarajaOikeudetSelection}
+            <th class="etp-table--th">
+              {i18n(i18nRoot + '.decision-order-notice-bailiff.district-court')}
+            </th>
+
+            <th class="etp-table--th">
+              {i18n(i18nRoot + '.decision-order-notice-bailiff.bailiff-email')}
+            </th>
+          {/if}
 
           <th class="etp-table--th">
             {i18n(i18nRoot + '.esikatselu')}
           </th>
+
+          {#if showCreateDocument}
+            <th class="etp-table--th">{i18n(i18nRoot + '.create-document')}</th>
+          {/if}
         </tr>
       </thead>
       <tbody class="etp-table--tbody">
@@ -94,27 +152,145 @@
                 - {Maybe.orSome('', osapuoli['rooli-description'])}
               {/if}
             </td>
-            <td
-              class="etp-table--td"
-              class:text-error={Maybe.isSome(osapuoli.type.errorKey(osapuoli))}
-              title={Maybe.fold(
-                '',
-                key =>
-                  i18n('valvonta.kaytto.osapuoli.toimitustapa-errors.' + key),
-                osapuoli.type.errorKey(osapuoli)
-              )}>
-              {#if Maybe.isSome(osapuoli.type.errorKey(osapuoli))}
-                <span class="font-icon">warning</span>
-              {/if}
-              {#if manuallyDeliverableToimenpide}
-                {i18n('valvonta.kaytto.toimenpide.manually-sent')}
-              {:else}
-                {toimitustapaLabel(osapuoli)}
-              {/if}
-              {#if Osapuolet.toimitustapa.other(osapuoli)}
-                - {Maybe.orSome('', osapuoli['toimitustapa-description'])}
-              {/if}
-            </td>
+
+            {#if showDeliveryMethod}
+              <td
+                class="etp-table--td"
+                class:text-error={Maybe.isSome(
+                  osapuoli.type.errorKey(osapuoli)
+                )}
+                title={Maybe.fold(
+                  '',
+                  key =>
+                    i18n('valvonta.kaytto.osapuoli.toimitustapa-errors.' + key),
+                  osapuoli.type.errorKey(osapuoli)
+                )}>
+                {#if Maybe.isSome(osapuoli.type.errorKey(osapuoli))}
+                  <span class="font-icon">warning</span>
+                {/if}
+                {#if manuallyDeliverableToimenpide}
+                  {i18n('valvonta.kaytto.toimenpide.manually-sent')}
+                {:else}
+                  {toimitustapaLabel(osapuoli)}
+                {/if}
+                {#if Osapuolet.toimitustapa.other(osapuoli)}
+                  - {Maybe.orSome('', osapuoli['toimitustapa-description'])}
+                {/if}
+              </td>
+            {/if}
+
+            {#if showHallintoOikeudetSelection}
+              <td class="etp-table--td court-container">
+                {#if Osapuolet.isOmistaja(osapuoli) && Toimenpiteet.documentExistsForOsapuoli(toimenpide, osapuoli.id, Osapuolet.getOsapuoliType(osapuoli))}
+                  <Select
+                    bind:model={toimenpide}
+                    lens={R.lensPath([
+                      'type-specific-data',
+                      'osapuoli-specific-data',
+                      osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                      'hallinto-oikeus-id'
+                    ])}
+                    modelToItem={Maybe.fold(
+                      Maybe.None(),
+                      Maybe.findById(R.__, hallintoOikeudet)
+                    )}
+                    itemToModel={Maybe.fold(Maybe.None(), it =>
+                      Maybe.Some(it.id)
+                    )}
+                    format={Maybe.fold(
+                      i18n('validation.no-selection'),
+                      Locales.label($locale)
+                    )}
+                    validators={R.path(
+                      [
+                        'type-specific-data',
+                        'osapuoli-specific-data',
+                        osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                        'hallinto-oikeus-id'
+                      ],
+                      schema
+                    )}
+                    items={Selects.addNoSelection(
+                      R.filter(isValid, hallintoOikeudet)
+                    )} />
+                {:else}
+                  {i18n('valvonta.kaytto.toimenpide.no-delivery')}
+                {/if}
+              </td>
+            {/if}
+
+            {#if showKarajaOikeudetSelection}
+              <td class="etp-table--td court-container">
+                {#if Osapuolet.isOmistaja(osapuoli) && Toimenpiteet.documentExistsForOsapuoli(toimenpide, osapuoli.id, Osapuolet.getOsapuoliType(osapuoli))}
+                  <Select
+                    bind:model={toimenpide}
+                    lens={R.lensPath([
+                      'type-specific-data',
+                      'osapuoli-specific-data',
+                      osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                      'karajaoikeus-id'
+                    ])}
+                    modelToItem={Maybe.fold(
+                      Maybe.None(),
+                      Maybe.findById(R.__, karajaoikeudet)
+                    )}
+                    itemToModel={Maybe.fold(Maybe.None(), it =>
+                      Maybe.Some(it.id)
+                    )}
+                    format={Maybe.fold(
+                      i18n('validation.no-selection'),
+                      Locales.label($locale)
+                    )}
+                    validators={R.path(
+                      [
+                        'type-specific-data',
+                        'osapuoli-specific-data',
+                        osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                        'karajaoikeus-id'
+                      ],
+                      schema
+                    )}
+                    items={Selects.addNoSelection(
+                      R.filter(isValid, karajaoikeudet)
+                    )} />
+                {:else}
+                  {i18n('valvonta.kaytto.toimenpide.no-delivery')}
+                {/if}
+              </td>
+              <td class="etp-table--td">
+                {#if Osapuolet.isOmistaja(osapuoli) && Toimenpiteet.documentExistsForOsapuoli(toimenpide, osapuoli.id, Osapuolet.getOsapuoliType(osapuoli))}
+                  <Input
+                    bind:model={toimenpide}
+                    required={true}
+                    type="email"
+                    lens={R.lensPath([
+                      'type-specific-data',
+                      'osapuoli-specific-data',
+                      osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                      'haastemies-email'
+                    ])}
+                    validators={R.path(
+                      [
+                        'type-specific-data',
+                        'osapuoli-specific-data',
+                        osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                        'haastemies-email'
+                      ],
+                      schema
+                    )}
+                    format={Maybe.orSome('')}
+                    parse={R.ifElse(
+                      R.isEmpty,
+                      R.always(Maybe.None()),
+                      Maybe.Some
+                    )}
+                    {i18n} />
+                {:else}
+                  {i18n('valvonta.kaytto.toimenpide.no-delivery')}
+                {/if}
+              </td>
+            {/if}
+
             <td class="etp-table--td">
               {#if Osapuolet.isOmistaja(osapuoli)}
                 {#if previewPending}
@@ -122,10 +298,14 @@
                     <Spinner smaller={true} />
                   </div>
                 {:else}
+                  <!-- TODO: Tämä pitää saada näkymään myös aina? -->
+
                   <div
                     class:text-primary={!disabled}
                     class:text-disabled={disabled}
                     class="cursor-pointer etp-table--td__center"
+                    role="button"
+                    class:invisible={isPreviewHidden(toimenpide, osapuoli)}
                     on:click|stopPropagation={disabled ||
                       preview(
                         osapuoli.type.preview(id, osapuoli.id, toimenpide)
@@ -141,6 +321,23 @@
                 {i18n(i18nRoot + '.fyi-disabled')}
               {/if}
             </td>
+
+            {#if showCreateDocument}
+              <td class="etp-table--td">
+                <div class="etp-table--td__center">
+                  {#if Osapuolet.isOmistaja(osapuoli)}
+                    <Checkbox
+                      bind:model={toimenpide}
+                      lens={R.lensPath([
+                        'type-specific-data',
+                        'osapuoli-specific-data',
+                        osapuoliSpecificDataIndexForOsapuoli(osapuoli),
+                        'document'
+                      ])} />
+                  {/if}
+                </div>
+              </td>
+            {/if}
           </tr>
         {/each}
       </tbody>
